uwcse403changemymoodv0/requirements.txt
Django==1.4.1
argparse==1.2.1
distribute==0.6.28
dj-database-url==0.2.1
django-piston==0.2.3
flickrapi==1.4.2
httplib2==0.7.6
psycopg2==2.4.5
python-gflags==2.0
wsgiref==0.1.2
